Move implementation to GenericValue::isActualVersionOf
