+++ BabelWiresLib/Types/Generic/genericValue.cpp
@@ -109,3 +109,72 @@ void babelwires::GenericValue::instantiate(const TypeSystem& typeSystem, const T
     // which would be updated if we encountered a nested generic type using exploration.
     m_wrappedValue = m_actualWrappedType.resolve(typeSystem).createValue(typeSystem);
 }
+
+bool babelwires::GenericValue::isActualVersionOf(const TypeRef& wrappedType) const {
+    struct Visitor {
+        Visitor(const GenericValue& genericValue, unsigned int level = 0)
+            : m_genericValue(genericValue)
+            , m_level(level) {}
+        bool operator()() { return false; } // Called when structurally different
+        bool operator()(std::monostate, std::monostate) { return true; }
+        bool operator()(const RegisteredTypeId& typeIdGen, const RegisteredTypeId& typeIdAct) {
+            // Simplifying restriction for now: registered types may not contain unbound type variables.
+            return typeIdGen == typeIdAct;
+        }
+        bool operator()(const TypeConstructorId& constructorIdRef,
+                        const TypeConstructorArguments& constructorArgumentsRef,
+                        const TypeConstructorId& constructorIdAct,
+                        const TypeConstructorArguments& constructorArgumentsAct) {
+            if (constructorIdRef != constructorIdAct) {
+                return false;
+            }
+            if (constructorIdRef == TypeVariableTypeConstructor::getThisIdentifier()) {
+                const TypeVariableData variableDataRef =
+                    TypeVariableTypeConstructor::extractValueArguments(constructorArgumentsRef.m_valueArguments);
+                const TypeVariableData variableDataAct =
+                    TypeVariableTypeConstructor::extractValueArguments(constructorArgumentsAct.m_valueArguments);
+                if ((variableDataRef.m_typeVariableIndex != variableDataAct.m_typeVariableIndex) ||
+                    (variableDataRef.m_numGenericTypeLevels != variableDataAct.m_numGenericTypeLevels)) {
+                    return false;
+                }
+                if (variableDataAct.m_numGenericTypeLevels == m_level) {
+                    assert(variableDataAct.m_typeVariableIndex <= m_genericValue.m_typeVariableAssignments.size());
+                    assert(!m_genericValue.m_typeVariableAssignments[variableDataAct.m_typeVariableIndex] ||
+                           (constructorArgumentsAct.m_typeArguments.size() == 1));
+                    assert(!m_genericValue.m_typeVariableAssignments[variableDataAct.m_typeVariableIndex] ||
+                           (m_genericValue.m_typeVariableAssignments[variableDataAct.m_typeVariableIndex] ==
+                            constructorArgumentsAct.m_typeArguments[0]));
+                    // Allowed to differ here.
+                    return true;
+                }
+            }
+            unsigned int level = m_level;
+            if (constructorIdRef == GenericTypeConstructor::getThisIdentifier()) {
+                // Variables under this constructor would have to jump over this GenericType to
+                // reach the one were processing.
+                ++level;
+            }
+            // Compare the type and value arguments.
+            if ((constructorArgumentsRef.m_typeArguments.size() != constructorArgumentsAct.m_typeArguments.size()) ||
+                (constructorArgumentsRef.m_valueArguments.size() != constructorArgumentsAct.m_valueArguments.size())) {
+                return false;
+            }
+            for (unsigned int i = 0; i < constructorArgumentsAct.m_valueArguments.size(); ++i) {
+                if (constructorArgumentsRef.m_valueArguments[i] != constructorArgumentsAct.m_valueArguments[i]) {
+                    return false;
+                }
+            }
+            for (unsigned int i = 0; i < constructorArgumentsAct.m_typeArguments.size(); ++i) {
+                Visitor childVisitor(m_genericValue, level);
+                if (!TypeRef::visit<Visitor, bool>(childVisitor, constructorArgumentsRef.m_typeArguments[i],
+                                                   constructorArgumentsAct.m_typeArguments[i])) {
+                    return false;
+                }
+            }
+            return true;
+        }
+        const GenericValue& m_genericValue;
+        unsigned int m_level;
+    } visitor(*this);
+    return TypeRef::visit<Visitor, bool>(visitor, wrappedType, m_actualWrappedType);
+}
